Move the controller files into a folder for cleaner project structure
