+++ .idea/libraries/xerial_sqlite_jdbc.xml
@@ -1,0 +1,10 @@
+<component name="libraryTable">
+  <library name="xerial.sqlite.jdbc" type="repository">
+    <properties maven-id="org.xerial:sqlite-jdbc:3.36.0.3" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/xerial/sqlite-jdbc/3.36.0.3/sqlite-jdbc-3.36.0.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
