+++ src/main/java/Application/DatabaseConnection.java
@@ -1,55 +1,28 @@
 package Application;
 
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.sql.DriverManager;
 
 public class DatabaseConnection {
-    private static DatabaseConnection instance;
-    private static HikariDataSource dataSource;
+    private static Connection instance = null;
 
     private DatabaseConnection() {
-        HikariConfig config = new HikariConfig();
-        config.setJdbcUrl("jdbc:mysql://localhost:3306/student_planner_db?serverTimezone=UTC");
-        config.setUsername("student_user");
-        config.setPassword("SecurePass123");
-        config.setMaximumPoolSize(10);
-        config.setMinimumIdle(2);
-        config.setIdleTimeout(30000);
-        config.setMaxLifetime(1800000);
-
-        dataSource = new HikariDataSource(config);
+        String url = "jdbc:sqlite:database.db";
+        try {
+            instance = DriverManager.getConnection(url);
+        } catch (SQLException sqlEx) {
+            System.err.println(sqlEx);
+        }
     }
 
-    public static synchronized DatabaseConnection getInstance() {
+    public static Connection getInstance() {
         if (instance == null) {
-            instance = new DatabaseConnection();
+            new DatabaseConnection();
         }
         return instance;
-    }
-
-    public Connection getConnection() throws SQLException {
-        return dataSource.getConnection();
-    }
-
-    public static void closeResources(Statement statement, ResultSet resultSet) {
-        try {
-            if (resultSet != null) resultSet.close();
-            if (statement != null) statement.close();
-        } catch (SQLException e) {
-            System.out.println(e);
-        }
-    }
-
-    public static void closeConnection(Connection connection) {
-        try {
-            if (connection != null) connection.close();
-        } catch (SQLException e) {
-            System.out.println(e);
-        }
     }
 }
