+++ src/main/java/module-info.java
@@ -2,9 +2,11 @@ module Application {
     requires javafx.controls;
     requires javafx.fxml;
     requires java.sql;
-    requires com.zaxxer.hikari;
 
     // Open package to JavaFX runtime
+    opens Application.Controllers to javafx.graphics, javafx.fxml;
+
     exports Application;
-    opens Application to javafx.graphics, javafx.fxml;
+    exports Application.Controllers;
+
 }
